fix: corrige erro de linting do React Fast Refresh e limpa warnings de coverage
