+++ frontend/src/contexts/SocketContext.tsx
@@ -1,4 +1,4 @@
-import React, { type ReactNode, useContext } from 'react';
+import React, { type ReactNode } from 'react';
 import { useSocket } from '../hooks/useSocket';
 import { SocketContext, type SocketContextType } from './SocketContextType';
 
@@ -14,13 +14,5 @@ export const SocketProvider: React.FC<SocketProviderProps> = ({ children }) => {
       {children}
     </SocketContext.Provider>
   );
-};
-
-export const useSocketContext = (): SocketContextType => {
-  const context = useContext(SocketContext);
-  if (!context) {
-    throw new Error('useSocketContext must be used within a SocketProvider');
-  }
-  return context;
 };
